Don't auto-enable ac-js2-mode
Following elisp conventions, code should not auto-enable
itself: it should be possible for a user to have a large number of
packages installed without all of them being turned on. Indeed, this is
the case with the Emacs standard library.

This change therefore removes the autoloaded `add-hook` code here in
favour of advising the user how to explicitly enable the mode.